feat: add `customResources` to sessionSigs
