+++ packages/types/src/lib/interfaces.ts
@@ -314,7 +314,7 @@ export interface JsonSignChainDataRequest {
 
 export interface JsonSignSessionKeyRequestV1
   extends Pick<LitActionSdkParams, 'jsParams'>,
-    Pick<LitActionSdkParams, 'litActionIpfsId'> {
+  Pick<LitActionSdkParams, 'litActionIpfsId'> {
   sessionKey: string;
   authMethods: AuthMethod[];
   pkpPublicKey?: string;
@@ -503,7 +503,7 @@ export interface JsonExecutionSdkParamsTargetNode
 
 export interface JsonExecutionSdkParams
   extends Pick<LitActionSdkParams, 'jsParams'>,
-    ExecuteJsAdvancedOptions {
+  ExecuteJsAdvancedOptions {
   /**
    *  JS code to run on the nodes
    */
@@ -579,7 +579,7 @@ export interface SessionSigsOrAuthSig {
 
 export interface DecryptRequestBase
   extends SessionSigsOrAuthSig,
-    MultipleAccessControlConditions {
+  MultipleAccessControlConditions {
   /**
    * The chain name of the chain that this contract is deployed on.  See LIT_CHAINS for currently supported chains.
    */
@@ -624,7 +624,7 @@ export interface EncryptFileRequest extends DecryptRequestBase {
   file: AcceptedFileType;
 }
 
-export interface DecryptRequest extends EncryptResponse, DecryptRequestBase {}
+export interface DecryptRequest extends EncryptResponse, DecryptRequestBase { }
 
 export interface DecryptResponse {
   // The decrypted data as a Uint8Array
@@ -646,10 +646,10 @@ export interface SigResponse {
 
 export interface ExecuteJsResponseBase {
   signatures:
-    | {
-        sig: SigResponse;
-      }
-    | any;
+  | {
+    sig: SigResponse;
+  }
+  | any;
 }
 
 /**
@@ -679,7 +679,7 @@ export interface ExecuteJsNoSigningResponse extends ExecuteJsResponseBase {
   logs: string;
 }
 
-export interface LitNodePromise {}
+export interface LitNodePromise { }
 
 export interface SendNodeCommand {
   url: string;
@@ -688,10 +688,10 @@ export interface SendNodeCommand {
 }
 export interface SigShare {
   sigType:
-    | 'BLS'
-    | 'K256'
-    | 'ECDSA_CAIT_SITH' // Legacy alias of K256
-    | 'EcdsaCaitSithP256';
+  | 'BLS'
+  | 'K256'
+  | 'ECDSA_CAIT_SITH' // Legacy alias of K256
+  | 'EcdsaCaitSithP256';
 
   signatureShare: string;
   shareIndex?: number;
@@ -1091,6 +1091,11 @@ export interface SignSessionKeyProp extends LitActionSdkParams {
    * A LIT resource ability is a combination of a LIT resource and a LIT ability.
    */
   resourceAbilityRequests?: LitResourceAbilityRequest[];
+
+  /**
+   * By default, it handles minNodeCount of responses
+   */
+  handleAllResponses?: boolean;
 }
 
 export interface SignSessionKeyResponse {
@@ -1159,7 +1164,7 @@ export interface CommonGetSessionSigsProps {
 
 export interface BaseProviderGetSessionSigsProps
   extends CommonGetSessionSigsProps,
-    LitActionSdkParams {
+  LitActionSdkParams {
   /**
    * This is a callback that will be used to generate an AuthSig within the session signatures. It's inclusion is required, as it defines the specific resources and abilities that will be allowed for the current session.
    */
@@ -1168,7 +1173,7 @@ export interface BaseProviderGetSessionSigsProps
 
 export interface GetSessionSigsProps
   extends CommonGetSessionSigsProps,
-    LitActionSdkParams {
+  LitActionSdkParams {
   /**
    * This is a callback that will be used to generate an AuthSig within the session signatures. It's inclusion is required, as it defines the specific resources and abilities that will be allowed for the current session.
    */
@@ -1692,7 +1697,7 @@ export interface LoginUrlParams {
   error: string | null;
 }
 
-export interface BaseAuthenticateOptions {}
+export interface BaseAuthenticateOptions { }
 
 export interface EthWalletAuthenticateOptions extends BaseAuthenticateOptions {
   /**
@@ -1758,9 +1763,9 @@ export interface MintCapacityCreditsPerKilosecond
 }
 export interface MintCapacityCreditsContext
   extends MintCapacityCreditsPerDay,
-    MintCapacityCreditsPerSecond,
-    MintCapacityCreditsPerKilosecond,
-    GasLimitParam {}
+  MintCapacityCreditsPerSecond,
+  MintCapacityCreditsPerKilosecond,
+  GasLimitParam { }
 export interface MintCapacityCreditsRes {
   rliTxHash: string;
   capacityTokenId: any;
@@ -1883,12 +1888,12 @@ export interface LitActionSdkParams {
    * An object that contains params to expose to the Lit Action.  These will be injected to the JS runtime before your code runs, so you can use any of these as normal variables in your Lit Action.
    */
   jsParams?:
-    | {
-        [key: string]: any;
-        publicKey?: string;
-        sigName?: string;
-      }
-    | any;
+  | {
+    [key: string]: any;
+    publicKey?: string;
+    sigName?: string;
+  }
+  | any;
 }
 
 export interface LitEndpoint {
@@ -1910,7 +1915,7 @@ export interface SignerLike {
 
 export interface GetPkpSessionSigs
   extends CommonGetSessionSigsProps,
-    LitActionSdkParams {
+  LitActionSdkParams {
   pkpPublicKey: string;
 
   /**
@@ -1924,6 +1929,8 @@ export interface GetPkpSessionSigs
   authMethods?: AuthMethod[];
 
   ipfsOptions?: IpfsOptions;
+
+  handleAllResponses?: boolean;
 }
 
 /**
@@ -1936,18 +1943,18 @@ export type GetLitActionSessionSigs = CommonGetSessionSigsProps &
   Pick<Required<LitActionSdkParams>, 'jsParams'> &
   (
     | (Pick<Required<LitActionSdkParams>, 'litActionCode'> & {
-        litActionIpfsId?: never;
-      })
+      litActionIpfsId?: never;
+    })
     | (Pick<Required<LitActionSdkParams>, 'litActionIpfsId'> & {
-        litActionCode?: never;
-      })
+      litActionCode?: never;
+    })
   ) & {
     ipfsOptions?: IpfsOptions;
 
     /**
      * Special property to return all node responses
      */
-    getNodeResponses?: boolean;
+    handleAllResponses?: boolean;
   };
 
 export interface SessionKeyCache {
