+++ packages/wrapped-keys/src/lib/api/utils.ts
@@ -108,31 +108,3 @@ export function getPkpAccessControlCondition(
     },
   };
 }
-
-export function parseCustoemrResourcesFromNodeResponses(responses: any) {
-  return ((responses.map((item: any) => {
-    return {
-      signedMessage: item.siweMessage
-    };
-  }).map((v: any) => {
-    const signedMessage = v.signedMessage;
-    const urnLine = signedMessage.match(/urn:recap:[\w\d]+/)![0];
-
-    const authContext = JSON.parse(atob(urnLine.split(':')[2])).att['lit-resolvedauthcontext://*']['Auth/Auth'][0]['auth_context'];
-
-    const extractedCustomAuthResource = (authContext['customAuthResource']).slice(8, -2);
-    const formattedCustomAuthResource = extractedCustomAuthResource.replace(/\\"/g, '"');
-    let result;
-
-    try {
-      result = JSON.parse(formattedCustomAuthResource);
-    } catch (e) {
-      result = extractedCustomAuthResource
-    }
-
-    return {
-      authContext,
-      formattedCustomAuthResource: result,
-    };
-  }).find((item: any) => item.formattedCustomAuthResource !== 'undefined')))?.formattedCustomAuthResource
-}
